src/requirements.txt
gunicorn==20.0.4
Flask==1.1.2
redis==3.5.2
rq==1.4.2

google-auth==1.16.0
google-api-python-client==1.8.4
google-cloud-tasks==1.5.0
